requirements.txt
flask
requests
llama-cpp-python